LDE-714: moved tests for append_cube
