+++ tests/io/cube/utils.py
@@ -1,9 +1,11 @@
+import math
 import pickle
 from collections import defaultdict
 
 import dask
 import dask.bag as db
 import dask.dataframe as ddf
+from pyarrow.parquet import ParquetFile
 
 
 def pickle_roundtrip(obj):
@@ -161,3 +163,21 @@ def wrap_bag_delete(f, blocksize):
         b.compute()
 
     return _driver
+
+
+def assert_num_row_groups(store, dataset, part_num_rows, part_chunk_size):
+    """
+    Assert that the row groups of each partition match the expectation based on the
+    number of rows and the chunk size
+    """
+    # Iterate over the partitions of each index value
+    for index, partitions in dataset.indices["p"].index_dct.items():
+        for part_key in partitions:
+            key = dataset.partitions[part_key].files["table"]
+            parquet_file = ParquetFile(store.open(key))
+            if part_chunk_size[index] is None:
+                assert parquet_file.num_row_groups == 1
+            else:
+                assert parquet_file.num_row_groups == math.ceil(
+                    part_num_rows[index] / part_chunk_size[index]
+                )
